+++ crates/machine/src/machine/machine_events.rs
@@ -1,5 +1,5 @@
 use super::Machine;
 
 pub struct MachineEvents {
-    pub on_next: Option<fn(&mut dyn Machine, &u64)>,
+    pub on_next: Option<fn(&mut dyn Machine, &u128)>,
 }
